pyproject.toml
[project]
name = "lerner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "mysql-connector-python>=9.3.0",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]
